Raise error if api_key in VGE not settled
